+++ RELEASE_NOTES.md
@@ -1,6 +1,22 @@
 Release Notes
 =============
 
+v1.0.0-alpha.x
+--------------
+
+### Breaking changes
+
+- Renamed the top-level python package to `openassetio_manager_bal`.
+  [#9](https://github.com/OpenAssetIO/OpenAssetIO-Manager-BAL/issues/9)
+
+### New features
+
+- BAL now exposes an `openassetio.manager_plugin` entry point, and can
+  be installed via `python -m pip` or similar. This allows it to be used
+  with OpenAssetIO without needing to extend `OPENASSETIO_PLUGIN_PATH`.
+  [#9](https://github.com/OpenAssetIO/OpenAssetIO-Manager-BAL/issues/9)
+
+
 v1.0.0-alpha.1
 --------------